move segment file
