+++ packages/backend-lib/tsconfig.jest.json
@@ -1,4 +1,6 @@
 {
   "extends": "./tsconfig.json",
   "files": ["../../node_modules/jest-expect-message/types/index.d.ts"],
+  "include": ["./test/**/*.ts"]
+
 }
