+++ packages/backend-lib/tsconfig.json
@@ -5,8 +5,8 @@
     "outDir": "dist",
     "skipLibCheck": true,
   },
-  "exclude": ["node_modules", "dist"],
-  "include": ["./src/**/*.ts", "./scripts/**/*.ts", "./test/**/*.ts"],
+  "exclude": ["node_modules", "dist", "**/*.test.ts", "**/*.test.tsx"],
+  "include": ["./src/**/*.ts", "./scripts/**/*.ts"],
   "references": [
     {
       "path": "../isomorphic-lib"
